baseline/requirements.txt
mock
numpy
pandas
scikit-learn
scipy
matplotlib
aiohttp
rasa
rasa_nlu
spacy
sklearn_crfsuite
tensorflow
pandas_ml
unicodecsv
